requirements.txt
Click==7.0
colorama==0.4.3
Flask==1.1.1
itsdangerous==1.1.0
Jinja2==2.11.1
MarkupSafe==1.1.1
mistune==0.8.4
pkg-resources==0.0.0
python-dotenv==0.11.0
PyYAML==5.3
selenium==3.141.0
urllib3==1.25.8
Werkzeug==1.0.0
